hm/sys/desktop/chat/msteams: Break out service option from autostart

+++ homeModules/sys/desktop/chat/msteams.nix
@@ -5,8 +5,10 @@ let
 in
 {
   options.sys.desktop.chat.msteams = {
-    enable = lib.mkEnableOption "Microsoft teams";
+    enable = lib.mkEnableOption "Microsoft Teams";
+
     package = lib.mkPackageOption pkgs "teams" { };
+    systemd-service = lib.mkEnableOption "ms-teams Systemd user service" // { default = true; };
     autostart = lib.mkEnableOption "Microsoft Teams on startup";
   };
 
@@ -15,7 +17,7 @@ in
 
     xdg.mimeApps.defaultApplications."x-scheme-handler/msteams" = "${cfg.package.name}.desktop";
 
-    systemd.user.services.msteams-client = lib.mkIf cfg.autostart {
+    systemd.user.services.msteams-client = lib.mkIf cfg.systemd-service {
       Unit = {
         Description = "${cfg.package.name} Microsoft Teams client";
         Requires = [
@@ -34,7 +36,7 @@ in
         Environment = "NIXOS_OZONE_WL=1";
       };
 
-      Install.WantedBy = [ "graphical-session.target" ];
+      Install.WantedBy = lib.mkIf cfg.autostart [ "graphical-session.target" ];
     };
 
     # For whatever reason, teams likes to overwrite the mimetypes, even if it's
@@ -46,10 +48,8 @@ in
           filename = "${config.xdg.configHome}/mimeapps.list";
         in
         lib.hm.dag.entryBefore [ "checkLinkTargets" ] ''
-          if [ -e ${filename} ]; then
-            if [ ! -L ${filename} ]; then
-              $DRY_RUN_CMD rm $VERBOSE_ARG ${filename}
-            fi
+          if [ -e ${filename} ] && [ ! -L ${filename} ]; then
+            $DRY_RUN_CMD rm $VERBOSE_ARG ${filename}
           fi
         ''
       );
